\name{cormotif internal}
\alias{modt.f0.loglike}
\alias{modt.f1.loglike}
\alias{cmfit}
\alias{cmfitall}
\alias{cmfitsep}
\alias{cmfitfull}
\alias{limmafit}
\alias{generatetype}
\title{Correlation Motif Internal functions}
\description{These functions are not part of the package application programming interface and are not recommended to be used by the users.}

\usage{
modt.f0.loglike
modt.f1.loglike
cmfit
cmfitall
cmfitsep
cmfitfull
limmafit
generatetype
}

\references{Ji, H., Wei, Y.,(2011) Correlation Motif. Unpublished}