\name{simudata2}
\alias{simudata2}
\alias{simu2_groupid}
\alias{simu2_compgroup}
\title{Example dataset for Cormotif}
\description{Here we present three files needed for the various Correlation Motif fit functions.}

\details{simudata2 are combined from four
studies sharing the same 3,000 genes, each having two experiment conditions
and three samples for each condition. simudata2 saves the expression values for all genes and all sample arrays on log2 scale; simu2_groupid prepares the group label for each sample; and simu2_compgroup describes the study design }

\references{Ji, H., Wei, Y.,(2011) Correlation Motif. Unpublished}